+++ src/api/v1/apps/orders/router.py
@@ -78,3 +78,44 @@ async def update_product(product_id: int,
     product_service = ProductService(session)
     
     return product_service.delete_product(product_id=product_id)
+
+# endpoint to acces the update product template
+@router.get("/products/update", response_class=HTMLResponse)
+async def get_update_product_template(request: Request,
+                        current_user: dict = Depends(get_current_user_from_cookie),
+                        session: Session = Depends(get_session)
+                        ) -> Response:
+    product_service = ProductService(session)
+    product_list = product_service.get_products()
+    return templates.TemplateResponse(
+        request=request,
+        name="products_update.html",
+        context={
+            "config": settings,
+            "products": product_list
+        }
+    )
+
+@router.get("/products/{product_id}",
+            response_model=ProductRead, 
+            status_code=status.HTTP_200_OK)
+async def get_product_by_id(
+    product_id: int,
+    request: Request,
+    current_user: dict = Depends(get_current_user_from_cookie),
+    session: Session = Depends(get_session)
+    ) -> Product | None:
+    """
+    Returns an individual product information
+    """
+
+    product_service = ProductService(session)
+    product = product_service.get_product_by_id(product_id=product_id)
+
+    if not product:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Product with ID index token {product_id} could not be located."
+        ) 
+
+    return product