+++ CHANGELOG.md
@@ -56,6 +56,8 @@ The bumping version is due to a new addition to save data.
 - `services.py` in `src/api/v1/apss/orders/`
 - Class `Products`in `src/api/v1/apss/orders/services.py`
 - endpoint POST orders/products in `src/api/v1/apss/orders/routes.py`
+- endpoint to retrieve product by id in `src/api/v1/apss/orders/routes.py`
+- template for update/delete product in `src/templates`
 
 ### Changed
 - N/A